test: Split StreamrClient.test across multiple files for maybe better perf.
