+++ test/integration/StreamrClient.test.ts
@@ -1,9 +1,8 @@
 import fs from 'fs'
 import path from 'path'
 
-import fetch from 'node-fetch'
-import { ControlLayer, MessageLayer } from 'streamr-client-protocol'
-import { wait, waitForEvent } from 'streamr-test-utils'
+import { MessageLayer } from 'streamr-client-protocol'
+import { wait } from 'streamr-test-utils'
 
 import { describeRepeats, uid, fakePrivateKey, getWaitForStorage, getPublishTestMessages, Msg } from '../utils'
 import { StreamrClient } from '../../src/StreamrClient'
@@ -14,12 +13,9 @@ import config from './config'
 import { Stream } from '../../src/stream'
 import { Subscription } from '../../src'
 
-const WebSocket = require('ws')
-
 const { StreamMessage } = MessageLayer
-const { SubscribeRequest, UnsubscribeRequest, ResendLastRequest } = ControlLayer
 
-const MAX_MESSAGES = 20
+const MAX_MESSAGES = 10
 
 describeRepeats('StreamrClient', () => {
     let expectErrors = 0 // check no errors by default
@@ -49,48 +45,6 @@ describeRepeats('StreamrClient', () => {
         c.onError = jest.fn()
         c.on('error', onError)
         return c
-    }
-
-    async function checkConnection() {
-        const c = createClient()
-        // create a temp client before connecting ws
-        // so client generates correct options.url for us
-        try {
-            await Promise.all([
-                Promise.race([
-                    fetch(c.options.restUrl),
-                    wait(1000).then(() => {
-                        throw new Error(`timed out connecting to: ${c.options.restUrl}`)
-                    })
-                ]),
-                Promise.race([
-                    new Promise((resolve, reject) => {
-                        const ws = new WebSocket(c.options.url)
-                        ws.once('open', () => {
-                            c.debug('open', c.options.url)
-                            resolve(undefined)
-                            ws.close()
-                        })
-                        ws.once('error', (err: any) => {
-                            c.debug('err', c.options.url, err)
-                            reject(err)
-                            ws.terminate()
-                        })
-                    }),
-                    wait(1000).then(() => {
-                        throw new Error(`timed out connecting to: ${c.options.url}`)
-                    })
-                ]),
-            ])
-        } catch (e) {
-            if (e.errno === 'ENOTFOUND' || e.errno === 'ECONNREFUSED') {
-                throw new Error('Integration testing requires that core-api '
-                    + 'and data-api ("entire stack") are running in the background. '
-                    + 'Instructions: https://github.com/streamr-dev/streamr-docker-dev#running')
-            } else {
-                throw e
-            }
-        }
     }
 
     beforeEach(() => {
@@ -125,505 +79,6 @@ describeRepeats('StreamrClient', () => {
         }
     })
 
-    describe('Connection', () => {
-        describe('bad config.url', () => {
-            it('emits error without autoconnect', async () => {
-                client = createClient({
-                    url: 'asdasd',
-                    autoConnect: false,
-                    autoDisconnect: false,
-                })
-
-                await expect(() => (
-                    client.connect()
-                )).rejects.toThrow()
-            })
-
-            it('rejects on connect without autoconnect', async () => {
-                client = createClient({
-                    url: 'asdasd',
-                    autoConnect: false,
-                    autoDisconnect: false,
-                })
-
-                await expect(() => (
-                    client.connect()
-                )).rejects.toThrow()
-            })
-
-            it('emits error with autoconnect after first call that triggers connect()', async () => {
-                expectErrors = 1
-
-                client = createClient({
-                    url: 'asdasd',
-                    autoConnect: true,
-                    autoDisconnect: true,
-                })
-                const client2 = createClient({
-                    autoConnect: true,
-                    autoDisconnect: true,
-                })
-
-                const otherOnError = jest.fn()
-                client2.on('error', otherOnError)
-
-                const stream = await client2.createStream({
-                    name: uid('stream')
-                }) // this will succeed because it uses restUrl config, not url
-
-                // publish should trigger connect
-                await expect(() => (
-                    client.publish(stream, {})
-                )).rejects.toThrow('Invalid URL')
-                // check error is emitted with same error before rejection
-                // not clear if emit or reject *should* occur first
-                expect(onError).toHaveBeenCalledTimes(1)
-                expect(client.onError).toHaveBeenCalledTimes(1)
-                expect(otherOnError).toHaveBeenCalledTimes(0)
-            }, 10000)
-        })
-
-        describe('bad config.restUrl', () => {
-            it('emits no error with no connection', async () => {
-                client = createClient({
-                    restUrl: 'asdasd',
-                    autoConnect: false,
-                    autoDisconnect: false,
-                })
-
-                await wait(100)
-            })
-
-            it('does not emit error with connect', async () => {
-                // error will come through when getting session
-                client = createClient({
-                    restUrl: 'asdasd',
-                    autoConnect: false,
-                    autoDisconnect: false,
-                })
-                await client.connect()
-                await wait(100)
-            })
-        })
-
-        describe('connect handling', () => {
-            it('connects the client', async () => {
-                client = createClient()
-                await client.connect()
-                expect(client.isConnected()).toBeTruthy()
-                // no error if already connected
-                await client.connect()
-                expect(client.isConnected()).toBeTruthy()
-                await client.disconnect()
-            })
-
-            it('does not error if connecting', async () => {
-                client = createClient()
-                const done = Defer()
-                client.connection.once('connecting', done.wrap(async () => {
-                    await client.connect()
-                    expect(client.isConnected()).toBeTruthy()
-                }))
-
-                await client.connect()
-                await done
-                expect(client.isConnected()).toBeTruthy()
-            })
-
-            it('connects if disconnecting', async () => {
-                const done = Defer()
-                client = createClient()
-                client.connection.once('disconnecting', done.wrap(async () => {
-                    await client.connect()
-                    expect(client.isConnected()).toBeTruthy()
-                    await client.disconnect()
-                }))
-
-                await client.connect()
-                await expect(async () => {
-                    await client.disconnect()
-                }).rejects.toThrow()
-                await done
-            })
-        })
-
-        describe('disconnect handling', () => {
-            it('disconnects the client', async () => {
-                client = createClient()
-                // no error if already disconnected
-                await client.disconnect()
-                await client.connect()
-                await client.disconnect()
-                expect(client.isDisconnected()).toBeTruthy()
-            })
-
-            it('does not error if disconnecting', async () => {
-                client = createClient()
-                const done = Defer()
-                client.connection.once('disconnecting', done.wrap(async () => {
-                    await client.disconnect()
-                    expect(client.isDisconnected()).toBeTruthy()
-                }))
-                await client.connect()
-                await client.disconnect()
-                await done
-            })
-
-            it('disconnects if connecting', async () => {
-                const done = Defer()
-                client = createClient()
-                client.connection.once('connecting', done.wrap(async () => {
-                    await client.disconnect()
-                    expect(client.isDisconnected()).toBeTruthy()
-                }))
-                await expect(async () => {
-                    await client.connect()
-                }).rejects.toThrow()
-                await done
-            })
-
-            it('does not reconnect after purposeful disconnect', async () => {
-                client = createClient()
-                await client.connect()
-                const done = Defer()
-
-                client.once('disconnected', done.wrap(async () => {
-                    await client.disconnect()
-                }))
-
-                client.connection.socket.close()
-                await done
-                await wait(2500)
-                expect(client.isDisconnected()).toBeTruthy()
-            })
-        })
-
-        describe('connect during disconnect', () => {
-            it('can connect after disconnect', async () => {
-                const done = Defer()
-                client = createClient()
-                client.once('connected', done.wrapError(async () => {
-                    await expect(async () => {
-                        await client.disconnect()
-                    }).rejects.toThrow()
-                }))
-                client.once('disconnected', done.wrapError(async () => {
-                    client.once('connected', done.wrapError(async () => {
-                        await client.disconnect()
-                        done.resolve(undefined)
-                    }))
-
-                    await expect(async () => {
-                        await client.connect()
-                    }).rejects.toThrow()
-                }))
-                await expect(async () => {
-                    await client.connect()
-                }).rejects.toThrow()
-                await done
-            })
-
-            it('can disconnect before connected', async () => {
-                client = createClient()
-
-                const t = expect(async () => {
-                    await client.connect()
-                }).rejects.toThrow()
-                await client.disconnect()
-                await t
-            })
-
-            it('can disconnect before connected', async () => {
-                client = createClient()
-                const t = expect(async () => {
-                    await client.connect()
-                }).rejects.toThrow()
-                await client.disconnect()
-                await t
-            })
-
-            it('can connect', async () => {
-                client = createClient()
-                const done = Defer()
-                await client.connect()
-
-                client.connection.once('disconnecting', done.wrap(async () => {
-                    await client.connect()
-                    await client.disconnect()
-                }))
-
-                await expect(async () => {
-                    await client.disconnect()
-                }).rejects.toThrow()
-                await done
-            })
-
-            it('can reconnect on unexpected close', async () => {
-                client = createClient()
-                await client.connect()
-
-                client.connection.socket.close()
-                expect(client.isConnected()).not.toBeTruthy()
-                await client.connection.nextConnection()
-                expect(client.isConnected()).toBeTruthy()
-            })
-
-            it('will resolve original disconnect', async () => {
-                const done = Defer()
-                client = createClient()
-
-                await client.connect()
-
-                client.connection.once('disconnecting', done.wrap(async () => {
-                    await client.connect()
-                }))
-                await expect(async () => {
-                    await client.disconnect()
-                }).rejects.toThrow()
-                await done
-            })
-
-            it('has connection state transitions in correct order', async () => {
-                client = createClient()
-                const done = Defer()
-                const connectionEventSpy = jest.spyOn(client.connection, 'emit')
-
-                await client.connect()
-
-                client.connection.once('disconnecting', done.wrap(async () => {
-                    await client.connect()
-                    const eventNames = connectionEventSpy.mock.calls.map(([eventName]) => eventName)
-                    expect(eventNames).toEqual([
-                        'connecting',
-                        'connected',
-                        'disconnecting',
-                    ])
-                    expect(client.isConnected()).toBeTruthy()
-                }))
-
-                await expect(async () => {
-                    await client.disconnect()
-                }).rejects.toThrow()
-                await done
-            }, 5000)
-
-            it('should not subscribe to unsubscribed streams on reconnect', async () => {
-                client = createClient()
-                await client.connect()
-                const sessionToken = await client.session.getSessionToken()!
-
-                const stream = await client.createStream({
-                    name: uid('stream')
-                })
-
-                const connectionEventSpy = jest.spyOn(client.connection, '_send')
-                const sub = await client.subscribe(stream.id, () => {})
-                await wait(100)
-                await client.unsubscribe(sub)
-                await client.disconnect()
-                await client.connect()
-                await client.disconnect()
-                // key exchange stream subscription should not have been sent yet
-                expect(connectionEventSpy.mock.calls).toHaveLength(2)
-
-                // check whole list of calls after reconnect and disconnect
-                expect(connectionEventSpy.mock.calls[0]).toEqual([new SubscribeRequest({
-                    streamId: stream.id,
-                    streamPartition: 0,
-                    sessionToken: sessionToken!,
-                    requestId: connectionEventSpy.mock.calls[0][0].requestId,
-                })])
-
-                expect(connectionEventSpy.mock.calls[1]).toEqual([new UnsubscribeRequest({
-                    streamId: stream.id,
-                    streamPartition: 0,
-                    // @ts-expect-error
-                    sessionToken: sessionToken!,
-                    requestId: connectionEventSpy.mock.calls[1][0].requestId,
-                })])
-            })
-
-            it('should not subscribe after resend() on reconnect', async () => {
-                client = createClient()
-                await client.connect()
-                const sessionToken = await client.session.getSessionToken()!
-
-                const stream = await client.createStream({
-                    name: uid('stream')
-                })
-
-                const connectionEventSpy = jest.spyOn(client.connection, 'send')
-                const sub = await client.resend({
-                    stream: stream.id,
-                    resend: {
-                        last: 10
-                    }
-                })
-
-                const msgs = await sub.collect()
-
-                await wait(2000)
-                await client.connection.socket.close()
-                await client.connection.nextConnection()
-
-                // check whole list of calls after reconnect and disconnect
-                expect(connectionEventSpy.mock.calls[0]).toEqual([new ResendLastRequest({
-                    streamId: stream.id,
-                    streamPartition: 0,
-                    sessionToken: sessionToken!,
-                    numberLast: 10,
-                    requestId: connectionEventSpy.mock.calls[0][0].requestId,
-                })])
-                expect(msgs).toEqual([])
-
-                // key exchange stream subscription should not have been sent yet
-                expect(connectionEventSpy.mock.calls.length).toEqual(1)
-                await client.disconnect()
-            }, 10000)
-
-            it('does not try to reconnect', async () => {
-                client = createClient()
-                await client.connect()
-
-                const onConnectAfterDisconnecting = Defer()
-                // should not try connecting after disconnect (or any other reason)
-                client.connection.once('disconnecting', onConnectAfterDisconnecting.wrap(async () => {
-                    await client.connect()
-                }))
-
-                await expect(async () => {
-                    await client.disconnect()
-                }).rejects.toThrow()
-                await onConnectAfterDisconnecting
-                expect(client.isConnected()).toBe(true)
-                await client.disconnect()
-                const onConnecting = jest.fn()
-                client.once('connecting', onConnecting)
-                // wait for possible reconnections
-                await wait(2000)
-                expect(onConnecting).toHaveBeenCalledTimes(0)
-                expect(client.isConnected()).toBe(false)
-            }, 10000)
-        })
-
-        describe('publish/subscribe connection handling', () => {
-            describe('publish', () => {
-                it('will connect if not connected if autoconnect set', async () => {
-                    const done = Defer()
-                    client = createClient({
-                        autoConnect: true,
-                        autoDisconnect: true,
-                    })
-
-                    const stream = await client.createStream({
-                        name: uid('stream')
-                    })
-                    expect(client.isDisconnected()).toBeTruthy()
-
-                    const message = {
-                        id2: uid('msg')
-                    }
-
-                    client.once('connected', done.wrap(() => {}))
-                    await client.publish(stream.id, message)
-                    await done
-                    // wait in case of delayed errors
-                    await wait(250)
-                })
-
-                it('errors if disconnected autoconnect set', async () => {
-                    client = createClient({
-                        autoConnect: true,
-                        autoDisconnect: true,
-                    })
-
-                    await client.connect()
-                    const stream = await client.createStream({
-                        name: uid('stream')
-                    })
-
-                    const message = {
-                        id1: uid('msg')
-                    }
-                    const p = client.publish(stream.id, message)
-                    await wait(0)
-                    await client.disconnect() // start async disconnect after publish started
-                    await expect(p).rejects.toThrow()
-                    expect(client.isDisconnected()).toBeTruthy()
-                    // wait in case of delayed errors
-                    await wait(250)
-                })
-
-                it('errors if disconnected autoconnect not set', async () => {
-                    client = createClient({
-                        autoConnect: false,
-                        autoDisconnect: true,
-                    })
-
-                    await client.connect()
-                    const stream = await client.createStream({
-                        name: uid('stream')
-                    })
-
-                    const message = {
-                        id1: uid('msg')
-                    }
-                    const p = client.publish(stream.id, message)
-                    await wait(0)
-                    client.debug('about to intentionally fail publish...')
-                    await client.disconnect() // start async disconnect after publish started
-                    await expect(p).rejects.toThrow('Failed to publish')
-                    expect(client.isDisconnected()).toBeTruthy()
-                    // wait in case of delayed errors
-                    await wait(500)
-                }, 10000)
-            })
-
-            describe('subscribe', () => {
-                it('does not error if disconnect after subscribe', async () => {
-                    client = createClient({
-                        autoConnect: true,
-                        autoDisconnect: true,
-                    })
-
-                    await client.connect()
-                    const stream = await client.createStream({
-                        name: uid('stream')
-                    })
-
-                    await client.subscribe({
-                        streamId: stream.id,
-                    }, () => {})
-
-                    await client.disconnect()
-                })
-
-                it('does not error if disconnect after subscribe with resend', async () => {
-                    client = createClient({
-                        autoConnect: true,
-                        autoDisconnect: true,
-                    })
-
-                    await client.connect()
-                    const stream = await client.createStream({
-                        name: uid('stream')
-                    })
-
-                    await client.subscribe({
-                        streamId: stream.id,
-                        resend: {
-                            from: {
-                                timestamp: 0,
-                            },
-                        },
-                    }, () => {})
-
-                    await client.disconnect()
-                })
-            })
-        })
-    })
-
     describe('StreamrClient', () => {
         let stream: Stream
         let waitForStorage: (...args: any[]) => Promise<void>
@@ -631,6 +86,7 @@ describeRepeats('StreamrClient', () => {
 
         // These tests will take time, especially on Travis
         const TIMEOUT = 30 * 1000
+        const WAIT_TIME = 600
 
         const attachSubListeners = (sub: Subscription) => {
             const onSubscribed = jest.fn()
@@ -709,21 +165,21 @@ describeRepeats('StreamrClient', () => {
                 await client.publish(stream.id, {
                     test: 'client.publish',
                 })
-                await wait(TIMEOUT * 0.2)
+                await wait(WAIT_TIME)
             }, TIMEOUT)
 
             it('Stream.publish does not error', async () => {
                 await stream.publish({
                     test: 'Stream.publish',
                 })
-                await wait(TIMEOUT * 0.2)
+                await wait(WAIT_TIME)
             }, TIMEOUT)
 
             it('client.publish with Stream object as arg', async () => {
                 await client.publish(stream, {
                     test: 'client.publish.Stream.object',
                 })
-                await wait(TIMEOUT * 0.2)
+                await wait(WAIT_TIME)
             }, TIMEOUT)
 
             describe('subscribe/unsubscribe', () => {
@@ -761,7 +217,7 @@ describeRepeats('StreamrClient', () => {
                     expect(client.getSubscriptions()).toHaveLength(0) // lost subscription immediately
                     await unsubTask
                     await subTask
-                    await wait(TIMEOUT * 0.2)
+                    await wait(WAIT_TIME)
                     expect(events.onResent).toHaveBeenCalledTimes(0)
                     expect(events.onSubscribed).toHaveBeenCalledTimes(0)
                     expect(events.onUnsubscribed).toHaveBeenCalledTimes(0)
@@ -787,7 +243,7 @@ describeRepeats('StreamrClient', () => {
                         subTask,
                     ])
                     expect(client.getSubscriptions()).toHaveLength(0) // lost subscription immediately
-                    await wait(TIMEOUT * 0.2)
+                    await wait(WAIT_TIME)
                     expect(events.onResent).toHaveBeenCalledTimes(0)
                     expect(events.onSubscribed).toHaveBeenCalledTimes(0)
                     expect(events.onUnsubscribed).toHaveBeenCalledTimes(1)
@@ -815,7 +271,7 @@ describeRepeats('StreamrClient', () => {
                         expect(client.getSubscriptions()).toHaveLength(0) // lost subscription immediately
                         await unsubTask
                         await subTask
-                        await wait(TIMEOUT * 0.2)
+                        await wait(WAIT_TIME * 2)
                         expect(events.onResent).toHaveBeenCalledTimes(0)
                         expect(events.onSubscribed).toHaveBeenCalledTimes(0)
                         expect(events.onUnsubscribed).toHaveBeenCalledTimes(0)
@@ -843,7 +299,7 @@ describeRepeats('StreamrClient', () => {
 
                         await unsubTask
                         await subTask
-                        await wait(TIMEOUT * 0.2)
+                        await wait(WAIT_TIME)
                         expect(events.onResent).toHaveBeenCalledTimes(0)
                         expect(events.onSubscribed).toHaveBeenCalledTimes(0)
                         expect(events.onUnsubscribed).toHaveBeenCalledTimes(0)
@@ -863,7 +319,7 @@ describeRepeats('StreamrClient', () => {
                     await stream.publish(Msg())
                     await t
                     expect(client.getSubscriptions()).toHaveLength(0) // lost subscription immediately
-                    await wait(TIMEOUT * 0.2)
+                    await wait(WAIT_TIME)
                     expect(events.onResent).toHaveBeenCalledTimes(0)
                     expect(events.onSubscribed).toHaveBeenCalledTimes(0)
                     expect(events.onUnsubscribed).toHaveBeenCalledTimes(1)
@@ -1052,78 +508,6 @@ describeRepeats('StreamrClient', () => {
                 expect(received).toEqual([...published.slice(-2), msg])
                 expect(client.getSubscriptions()).toHaveLength(0)
             }, TIMEOUT)
-        })
-
-        describe('resend', () => {
-            let timestamps: number[] = []
-            let published: any[] = []
-
-            beforeEach(async () => {
-                publishTestMessages = getPublishTestMessages(client, {
-                    stream,
-                    waitForLast: true,
-                    waitForLastTimeout: 9000,
-                })
-
-                const publishedRaw = await publishTestMessages.raw(5)
-                timestamps = publishedRaw.map(([, raw]: any) => raw.streamMessage.getTimestamp())
-                published = publishedRaw.map(([msg]: any) => msg)
-            })
-
-            it('resend last', async () => {
-                const sub = await client.resend({
-                    stream: stream.id,
-                    resend: {
-                        last: 3,
-                    },
-                })
-
-                expect(await sub.collect()).toEqual(published.slice(-3))
-            })
-
-            it('resend from', async () => {
-                const sub = await client.resend({
-                    stream: stream.id,
-                    resend: {
-                        from: {
-                            timestamp: timestamps[3],
-                        },
-                    },
-                })
-
-                expect(await sub.collect()).toEqual(published.slice(3))
-            })
-
-            it('resend range', async () => {
-                const sub = await client.resend({
-                    stream: stream.id,
-                    resend: {
-                        from: {
-                            timestamp: timestamps[0],
-                        },
-                        to: {
-                            timestamp: timestamps[3] - 1,
-                        },
-                    },
-                })
-
-                expect(await sub.collect()).toEqual(published.slice(0, 3))
-            })
-
-            it('works with message handler + resent event', async () => {
-                const messages: any[] = []
-                const sub = await client.resend({
-                    stream: stream.id,
-                    resend: {
-                        last: 3,
-                    },
-                }, (msg) => {
-                    messages.push(msg)
-                })
-
-                await waitForEvent(sub, 'resent')
-                expect(messages).toEqual(published.slice(-3))
-            })
         })
 
         describe('utf-8 encoding', () => {
